backend/pyproject.toml
[project]
name = "ai-docanalyser2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "groq>=1.0.0",
    "matplotlib>=3.10.8",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "plotly>=6.5.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.53.1",
    "uvicorn>=0.40.0",
    "xlrd>=2.0.2",
]
